rename the metDatasets file
even though its an acronym, it is less confusing to have the name be lowercase since it is not a component
